Refactor: 채팅 쪽 비즈니스로직 정리 (#110)
